requirements.txt
telethon==1.36.0
python-dotenv>=1.0.1
openai>=1.57.0
httpx>=0.27.0
pydantic>=2.7.0
tzdata>=2024.1

